requirements.txt
streamlit
requests
opencv-python-headless
Pillow
numpy
python-dotenv